factored out the ASM code for aarch64
